Refactored project service to move out methods which are not core to its funtionality
