+++ lib/declarations.d.ts
@@ -149,7 +149,6 @@ declare module Project {
 		createProjectFileFromExistingProject(projectType: number): IFuture<void>;
 		createProjectFile(projectDir: string, projectType: number, properties: any): IFuture<void>;
 		createTemplateFolder(projectDir: string): IFuture<any>;
-		updateProjectProperty(projectData: any, mode: string, property: string, newValue: any, propSchema: any, useMapping?: boolean) : IFuture<void>;
 		getTempDir(extraSubdir?:string): IFuture<string>;
 	}
 
@@ -209,6 +208,8 @@ interface IProjectData {
 interface IProjectPropertiesService {
 	getProjectProperties(projectFile: string, isJsonProjectFile: boolean): IFuture<IProjectData>;
 	completeProjectProperties(properties: any): boolean;
+	updateProjectProperty(projectData: any, mode: string, property: string, newValue: any, propSchema: any, useMapping?: boolean) : IFuture<void>;
+	normalizePropertyName(property: string, schema: any): string;
 }
 
 interface IServerConfigurationData {
