+++ lib/services/project-properties-service.ts
@@ -1,7 +1,7 @@
 ///<reference path="../.d.ts"/>
-
 "use strict";
 
+import os = require("os");
 import xml2js = require("xml2js");
 import util = require("util");
 import Future = require("fibers/future");
@@ -10,6 +10,8 @@ import MobileHelper = require("../common/mobile/mobile-helper");
 
 export class ProjectPropertiesService implements IProjectPropertiesService {
 	constructor(private $fs: IFileSystem,
+				private $errors: IErrors,
+				private $injector: IInjector,
 				private $resources: IResourceLoader,
 				private $projectTypes: IProjectTypes) {
 	}
@@ -25,42 +27,6 @@ export class ProjectPropertiesService implements IProjectPropertiesService {
 
 			return properties;
 		}).future<IProjectData>()();
-	}
-
-	private getProjectPropertiesFromXmlProjectFile(projectFile: string): IFuture<any> {
-		return ((): any => {
-			var properties: any = {};
-
-			var parser = new xml2js.Parser();
-			var contents = this.$fs.readText(projectFile).wait();
-
-			var parseString = Future.wrap((str, callback) => {
-				return parser.parseString(str, callback);
-			});
-
-			var result: any = parseString(contents).wait();
-			var propertyGroup: any = result.Project.PropertyGroup[0];
-
-			var projectSchema = helpers.getProjectFileSchema(this.$projectTypes.Cordova).wait();
-			_.sortBy(Object.keys(projectSchema), key => key === "FrameworkVersion" ? -1 : 1).forEach((propertyName) => {
-				if (propertyGroup.hasOwnProperty(propertyName)) {
-					properties[propertyName] = propertyGroup[propertyName][0];
-
-					if (projectSchema[propertyName].flags) {
-						properties[propertyName] = properties[propertyName] !== "" ? properties[propertyName].split(";") : [];
-					}
-				}
-			});
-
-			// only old style .proj files (before project unification) have ProjectName
-			if (propertyGroup.ProjectName) {
-				properties.ProjectName = propertyGroup.ProjectName[0];
-			} else {
-				properties = null;
-			}
-
-			return properties;
-		}).future<any>()();
 	}
 
 	public completeProjectProperties(properties: any): boolean {
@@ -105,6 +71,205 @@ export class ProjectPropertiesService implements IProjectPropertiesService {
 		});
 
 		return updated;
+	}
+
+	public updateProjectProperty(projectData: any, mode: string, property: string, newValue: any, propSchema: any, useMapping: boolean = true) : IFuture<void> {
+		return ((): any => {
+			property = this.normalizePropertyName(property, propSchema);
+			var propData = propSchema[property];
+
+			var validate = (condition: boolean, ...args: string[]) => {
+				if(condition) {
+					if(propData.validationMessage) {
+						this.$errors.fail(propData.validationMessage);
+					} else {
+						this.$errors.fail.apply(this.$errors, _.rest(args, 0));
+					}
+				}
+			};
+
+			if (!propData) {
+				this.$errors.fail("Unrecognized project property '%s'", property);
+			}
+
+			if (!propData.flags) {
+				if (newValue.length !== 1) {
+					this.$errors.fail("Property '%s' is not a collection of flags. Specify only a single property value.", property);
+				}
+				if (mode === "add" || mode === "del") {
+					this.$errors.fail("Property '%s' is not a collection of flags. Use prop-set to set a property value.", property);
+				}
+			} else {
+				newValue = _.flatten(_.map(newValue, (value: string)  => value.split(";")));
+			}
+
+			var range = this.getPropRange(propData).wait();
+			if (range) {
+				newValue = _.map(newValue, (value: string) => value.toLowerCase());
+
+				var validValues: any;
+				if (_.isArray(range)) {
+					validValues = helpers.toHash(range, (value) => value.toLowerCase(), _.identity);
+				} else {
+					var keySelector = (value: any, key: string) => {
+						var result: string;
+						if (useMapping && value.input) {
+							result = value.input;
+						} else {
+							result = key;
+						}
+
+						return result.toLowerCase();
+					};
+
+					validValues = helpers.toHash(range, keySelector, (value, key) => key);
+				}
+
+				var badValues = _.reject(newValue, (value) => validValues[value]);
+
+				validate(badValues.length > 0, "Invalid property value%s for property '%s': '%s'", badValues.length > 1 ? "s" : "", property, badValues.join("; "));
+
+				newValue = _.map(newValue, (value) => validValues[value]);
+			}
+
+			if (!propData.flags) {
+				newValue = newValue[0];
+
+				if (propData.regex) {
+					var matchRegex = new RegExp(propData.regex);
+					validate(!matchRegex.test(newValue), "Value '%s' is not in the format expected by property %s. Expected to match /%s/", newValue, property, propData.regex);
+				}
+
+				if (propData.validator) {
+					var validator = this.$injector.resolve(propData.validator);
+					validator.validate(newValue);
+				}
+			}
+
+			var propertyValue = projectData[property];
+			if (propData.flags && _.isString(propertyValue)) {
+				propertyValue = propertyValue.split(";");
+			}
+
+			if (mode === "set") {
+				propertyValue = newValue;
+			} else if (mode === "del") {
+				propertyValue = _.difference(propertyValue, newValue);
+			} else if (mode === "add") {
+				propertyValue = _.union(propertyValue, newValue);
+			} else {
+				this.$errors.fail("Unknown property update mode '%s'", mode);
+			}
+
+			if (propertyValue.sort) {
+				propertyValue.sort();
+			}
+
+			if (propData.onChanging) {
+				this.$injector.dynamicCall(propData.onChanging, [propertyValue]).wait();
+			}
+
+			projectData[property] = propertyValue;
+		}).future<void>()();
+	}
+
+	public getProjectSchemaHelp(): IFuture<string> {
+		return (() => {
+			var result: string[] = [];
+			var schema = helpers.getProjectFilePartSchema(this.$projectTypes[this.$projectTypes.Cordova]).wait();
+			var title = util.format("Project properties for %s projects:", this.$projectTypes[this.$projectTypes.Cordova]);
+			result.push(this.getProjectSchemaPartHelp(schema, title));
+
+			schema = helpers.getProjectFilePartSchema(this.$projectTypes[this.$projectTypes.NativeScript]).wait();
+			title = util.format("Project properties for %s projects:", this.$projectTypes[this.$projectTypes.NativeScript]);
+			result.push(this.getProjectSchemaPartHelp(schema, title));
+
+			schema = helpers.getProjectFilePartSchema(this.$projectTypes[this.$projectTypes.Common]).wait();
+			title = "Common properties for all projects:";
+			result.push(this.getProjectSchemaPartHelp(schema, title));
+
+			return result.join(os.EOL + os.EOL);
+		}).future<string>()();
+	}
+
+	public normalizePropertyName(property: string, schema: any): string {
+		if (!property) {
+			return property;
+		}
+
+		var propLookup = helpers.toHash(schema, (value, key) => key.toLowerCase(), (value, key) => key);
+		return propLookup[property.toLowerCase()] || property;
+	}
+
+	private getProjectSchemaPartHelp(schema: string, title: string): string {
+		var help = [title];
+		_.each(schema, (value: any, key: any) => {
+			help.push(util.format("  %s - %s", key, value.description));
+			var range = this.getPropRange(value).wait();
+			if (range) {
+				help.push("    Valid values:");
+				_.each(range, (rangeDesc:any, rangeKey:any) => {
+					var desc = "      " + (_.isArray(range) ? rangeDesc : rangeDesc.input || rangeKey);
+					if (rangeDesc.description) {
+						desc += " - " + rangeDesc.description;
+					}
+					help.push(desc);
+				});
+			}
+			if (value.validationMessage) {
+				help.push("    " + value.validationMessage.replace(os.EOL, os.EOL + "    "));
+			}
+			else if (value.regex) {
+				help.push("    Valid values match /" + value.regex.toString() + "/");
+			}
+		});
+
+		return help.join(os.EOL);
+	}
+
+	private getPropRange(propData: any): IFuture<string[]>{
+		return (() => {
+			if (propData.dynamicRange) {
+				return this.$injector.dynamicCall(propData.dynamicRange).wait();
+			}
+			return propData.range;
+		}).future<string[]>()();
+	}
+
+	private getProjectPropertiesFromXmlProjectFile(projectFile: string): IFuture<any> {
+		return ((): any => {
+			var properties: any = {};
+
+			var parser = new xml2js.Parser();
+			var contents = this.$fs.readText(projectFile).wait();
+
+			var parseString = Future.wrap((str, callback) => {
+				return parser.parseString(str, callback);
+			});
+
+			var result: any = parseString(contents).wait();
+			var propertyGroup: any = result.Project.PropertyGroup[0];
+
+			var projectSchema = helpers.getProjectFileSchema(this.$projectTypes.Cordova).wait();
+			_.sortBy(Object.keys(projectSchema), key => key === "FrameworkVersion" ? -1 : 1).forEach((propertyName) => {
+				if (propertyGroup.hasOwnProperty(propertyName)) {
+					properties[propertyName] = propertyGroup[propertyName][0];
+
+					if (projectSchema[propertyName].flags) {
+						properties[propertyName] = properties[propertyName] !== "" ? properties[propertyName].split(";") : [];
+					}
+				}
+			});
+
+			// only old style .proj files (before project unification) have ProjectName
+			if (propertyGroup.ProjectName) {
+				properties.ProjectName = propertyGroup.ProjectName[0];
+			} else {
+				properties = null;
+			}
+
+			return properties;
+		}).future<any>()();
 	}
 }
 $injector.register("projectPropertiesService", ProjectPropertiesService);
